June2

+++ backend/routes/api/notebook/DELETEnotebook.js
@@ -13,13 +13,20 @@ const note = require("../../../db/models/note");
 
 router.delete("/profile/notebook", async (req, res) => {
   const notebookId = req.body.id;
+
+  if (notebookId === 0) {
+    return res.send({ length: 0 });
+  }
   const notebook = await db.NoteBook.findByPk(notebookId);
   const notes = await db.Note.findAll({ where: { notebookId } });
+
+  if (!notebook || !notes) {
+    return res.send({ length: 0 });
+  }
   const tagXnotes = await db.TagCrossNote.findAll({
     where: { noteId: notebookId },
   });
 
-  // console.log(tagXnotes, `this the cross table`);
   for (let i = 0; i < tagXnotes.length; i++) {
     let currentTXN = tagXnotes[i];
     await db.TagCrossNote.destroy({ where: { tagId: currentTXN.tagId } });
@@ -27,13 +34,9 @@ router.delete("/profile/notebook", async (req, res) => {
     await tags.destroy();
   }
 
-  //   console.log(`Does it come out here?
-  // `);
-
   for (let i = 0; i < notes.length; i++) {
     let currentNote = notes[i];
     await currentNote.destroy();
-    console.log(currentNote, `THE CURRENT NOTE`);
   }
 
   await notebook.destroy();
